backend_api/requirements/base.txt
Django==4.2.7
django-environ==0.11.2
psycopg==3.1.12
djangorestframework==3.14.0
django-filter==23.3
Pillow==10.1.0
djangorestframework-simplejwt==5.3.0 
drf-nested-routers==0.93.4